final heroku I hope so

+++ client/src/components/Person.js
@@ -1,19 +1,24 @@
 import React from "react";
+import classes from "./Person.module.css";
 import { Link } from "react-router-dom";
 const Person = (props) => {
   return (
     <tr>
-      <td>{props.people.name}</td>
+      <td >{props.people.name}</td>
       <td>{props.people.age}</td>
       <td>{props.people.gender}</td>
       <td>{props.people.date.substring(0, 10)}</td>
       <td>
-        <Link to={"/people/" + props.people._id}>edit</Link> |
+        <Link to={"/people/" + props.people._id}  className={` ${classes.actions} ${classes.border} btn btn-primary `}>edit</Link>
         <a
           href="#"
           onClick={() => {
-            props.deletePerson(props.people._id);
-          }}
+            if (window.confirm("Are you sure to delete?")) {
+              props.deletePerson(props.people._id)
+            } else {
+              window.alert = "You did not want to delete";
+            }
+          }}  className={`btn btn-danger ${classes.border}`}
         >
           delete
         </a>
